pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag"
version = "0.1.0"
description = "RAG ingestion and retrieval pipeline for Physical AI textbook"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.7.0",
    "google-genai>=1.0.0",
    "python-frontmatter>=1.0.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0",
    "click>=8.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "asyncpg>=0.29",
    "sse-starlette>=1.8",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
    "pytest-cov>=4.0",
    "httpx>=0.27",
]

[project.scripts]
rag = "rag.cli:main"

[tool.setuptools.packages.find]
include = ["rag*", "backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
